Refactor persistence manager

+++ src/storage/catalog/new_catalog_impl.cpp
@@ -449,17 +449,20 @@ std::unique_ptr<SystemCache> NewCatalog::RestoreCatalogCache(Storage *storage_pt
     // std::string meta_tree_str = meta_tree->ToJson().dump(4);
     // LOG_INFO(meta_tree_str);
     Config *config_ptr = storage_ptr->config();
-    std::string persistence_dir = config_ptr->PersistenceDir();
-    std::vector<std::filesystem::path> uuid_files;
-    std::regex uuid_pattern("^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$");
-    for (const auto &entry : std::filesystem::directory_iterator(persistence_dir)) {
-        if (entry.is_regular_file()) {
-            std::string filename = entry.path().filename().string();
-            if (std::regex_match(filename, uuid_pattern)) {
-                std::string file_name = entry.path().filename().string();
-                if (!meta_tree->pm_object_map_.contains(filename)) {
-                    // the pm object are redundant.
-                    VirtualStore::DeleteFile(entry.path().string());
+    PersistenceManager *pm = storage_ptr->persistence_manager();
+    if (pm != nullptr) {
+        std::string persistence_dir = config_ptr->PersistenceDir();
+        std::vector<std::filesystem::path> uuid_files;
+        std::regex uuid_pattern("^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$");
+        for (const auto &entry : std::filesystem::directory_iterator(persistence_dir)) {
+            if (entry.is_regular_file()) {
+                std::string filename = entry.path().filename().string();
+                if (std::regex_match(filename, uuid_pattern)) {
+                    std::string file_name = entry.path().filename().string();
+                    if (!meta_tree->pm_object_map_.contains(filename)) {
+                        // the pm object are redundant.
+                        VirtualStore::DeleteFile(entry.path().string());
+                    }
                 }
             }
         }
